requirements.txt
matplotlib==3.8.2
pandas==2.1.4
PyQt5==5.15.10
PyQtWebEngine==5.15.6
openpyxl==3.1.2
python-dotenv==1.0.0
SQLAlchemy==2.0.24